+++ packages/tswebextension/CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version changes -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## Unreleased
+
+### Changed
+
+- Added new field to MV3 config: `quickFixesRules` which will be applied to
+  the dynamic rules.
+
+
 ## [2.2.0-alpha.1] - 2024-08-30
 
 ### Changed